+++ lib/src/commands/commands.dart
@@ -1,3 +1,4 @@
+export 'build_commands/build_command.dart';
 export 'doctor_command.dart';
 export 'flutter_switch_command.dart';
 export 'install_commands/install_command.dart';