+++ test/src/commands/build_command_test.dart
@@ -1,0 +1,37 @@
+import 'package:flutter_compile/src/command_runner.dart';
+import 'package:flutter_compile/src/version.dart';
+import 'package:mason_logger/mason_logger.dart';
+import 'package:mocktail/mocktail.dart';
+import 'package:pub_updater/pub_updater.dart';
+import 'package:test/test.dart';
+
+class _MockLogger extends Mock implements Logger {}
+
+class _MockPubUpdater extends Mock implements PubUpdater {}
+
+void main() {
+  group('build', () {
+    late Logger logger;
+    late PubUpdater pubUpdater;
+    late FlutterCompileCommandRunner commandRunner;
+
+    setUp(() {
+      logger = _MockLogger();
+      pubUpdater = _MockPubUpdater();
+
+      when(
+        () => pubUpdater.getLatestVersion(any()),
+      ).thenAnswer((_) async => packageVersion);
+
+      commandRunner = FlutterCompileCommandRunner(
+        logger: logger,
+        pubUpdater: pubUpdater,
+      );
+    });
+
+    test('shows usage when no subcommand given', () async {
+      final result = await commandRunner.run(['build']);
+      expect(result, equals(ExitCode.usage.code));
+    });
+  });
+}